# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A percentile calculator with CLI and HTTP API",
        "title": "Outlier API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/wingnut128/outlier-go",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/calculate": {
            "post": {
                "description": "Calculate a percentile from an array of numeric values using linear interpolation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculate"
                ],
                "summary": "Calculate percentile from values",
                "parameters": [
                    {
                        "description": "Calculate Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/calculate/file": {
            "post": {
                "description": "Upload a JSON or CSV file and calculate percentile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculate"
                ],
                "summary": "Calculate percentile from file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Data file (JSON or CSV)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Percentile to calculate (default: 95)",
                        "name": "percentile",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is healthy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CalculateRequest": {
            "type": "object",
            "required": [
                "values"
            ],
            "properties": {
                "percentile": {
                    "type": "number"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "api.CalculateResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "percentile": {
                    "type": "number"
                },
                "result": {
                    "type": "number"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.HealthResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}